// SingInChineese/src/interfaces/enum.ts
export enum ResponseCode {
	success = 200,
	created = 201,
	badRequest = 422,
}